Support karydia config [resolves #71]
With deployed karydia config it is possible to set a cluster-wide karydia default behavior which is used if there is no specified setting (e.g. additional annotation) for a namespace found. With this, karydia is able to provide a configurable "secure by default" approach.

* Provide karydia config custom resource definition (CRD)

It is now possible to apply
- 'manifests/crd-config.yml' which registers karydia config CRD at your cluster
- 'manifests/config.yml' which registers the cluster-wide karydia default configuration to be used by karydia as default if there is no other namespace setting
Additionally, generated helpers for karydia config CRD are added.

* Initial karydia config import on start-up

With binary call parameter 'config-custom-resource' it is possible to change the name of the custom resource where to load the cluster-wide default configuration from for karydia.

* Consider karydia config
for automountServiceAccountToken and seccompProfile

* Adjust copyright headers

for touched files as discussed at #89