Distinguish between mutating and validating service token pod webhook (#93)
Option karydia.gardener.cloud/automountServiceAccountToken=remove-default
is only applicable for the mutating webhook. If the mutating hook
is not registered at the api server, but the option remove-default
is set, karydia will return a meaningful error message and will deny
the deployment. If karydia is configured as mutating and validating
webhook, karydia acts accordingly and patches the object.