[Test] Solve race-condition with default service account
Since the default service account is created with a varying delay
after a new nampspace is created, there might appear race-
conditions. For example when karydia annotations on the namespace
are updated right after namepsace creation, the consequences and
applied sercurity options for the automatically created default
service account vary depending on the reconcile timing!