[WORKAROUND] #183
This workaround is only valid till a consistent solution for issue #183 is found.
It adds the following annotations to the 'kube-system' namespace for deactivating the 'seccompProfile' and 'podSecurityContext' admission features (https://github.com/karydia/karydia/blob/master/docs/features.md#karydia-admission) in that namespace. Additionally, it adds one annotation for being able to identify 'kube-system' namespaces which used this workaround for cleaning things up later on.
- 'karydia.gardener.cloud/workaroundForIssue=183'
- 'karydia.gardener.cloud/seccompProfile=unconfined'
- 'karydia.gardener.cloud/podSecurityContext=none'